feat: disable zoom in button when location denied
